Update OpenStack project mappings
The list of which projects are part of OpenStack and which are not is
maintained here:

  http://git.openstack.org/cgit/openstack/governance/tree/reference/projects.yaml

Repos listed there are "OpenStack" - repos that are not are not. All of the
things listed in there are considered to be part of "One Project", not a
collection of independent projects.

The hints file has been updated to reflect this.

There were also projects that are OpenStack related but not associated
with OpenStack's official developer community. Those have been updated
to indicate that they are individual efforts (for instance, the repos in
openstack-charmers is an independent effort and should be tracked as
such)

Will do a follow up with some BigQuery files to cover getting the git
and other data from Gerrit. (OpenStack doesn't use PRs, so numbers
tracked from BigQuery involving those thing are currently inaccurate
-but the data does exist, so we can get it up to date)

It's possible there are other files that should be updated with the
updated hints file, but I wasn't sure what was "right" there.