Make the update more complete
In order to get all the data needed by the system, we need to pull the
full list of repos in OpenStack's gerrit, then cross-reference that with
the governance list. We then need to black-list a set of specific repos
that are sandboxes and governance repos and the like.

There are also some projects that are in the data set in BigQuery but
that are no longer actually repos in OpenStack due to renames. Put them
in.

While we're at it, go ahead and keep OpenStack Infra as its own category, but
don't skip it. It's an interesting Open Source project in its own right.

Including the script that generated the data. I know this is a ruby
repo and the script is in python. It's here as a placeholder/doc.
I'll get it translated to Ruby.